+++ src/main/java/com/evenly/took/feature/auth/client/google/GoogleUserInfoProvider.java
@@ -7,7 +7,8 @@ import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
 
-import com.evenly.took.feature.auth.client.google.dto.GoogleUserInfoResponse;
+import com.evenly.took.feature.auth.client.google.dto.response.GoogleUserInfoResponse;
+import com.evenly.took.feature.auth.client.google.error.GoogleUserInfoProviderErrorHandler;
 import com.evenly.took.global.config.properties.auth.GoogleUrlProperties;
 
 @Component
